requirements.txt
antlr4-python3-runtime==4.9.3
asciitree==0.3.3
braceexpand==0.1.7
certifi==2024.8.30
charset-normalizer==3.4.0
fast-jl==0.1.3
fasteners==0.19
filelock==3.16.1
fsspec==2024.10.0
ftfy==6.3.1
huggingface-hub==0.26.5
hydra-core==1.3.2
idna==3.10
jinja2==3.1.4
markupsafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numcodecs==0.14.1
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
open-clip-torch==2.29.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
setuptools==75.6.0
six==1.17.0
sympy==1.13.1
timm==1.0.12
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
-e file:///git/pdpl/training
traker==0.3.2
triton==3.1.0
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
webdataset==0.2.100
zarr==2.18.3
